+++ packages/components/inputs/date-time-input/src/date-time-input.spec.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { render, fireEvent } from '../../../../test/test-utils';
+import { render, fireEvent } from '../../../../../test/test-utils';
 import DateTimeInput from './date-time-input';
 
 // This component is used to enable easy testing.
