+++ packages/components/inputs/date-time-input/src/date-time-input.story.js
@@ -2,9 +2,9 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { withKnobs, select, boolean, text } from '@storybook/addon-knobs/react';
 import { action } from '@storybook/addon-actions';
-import Readme from './README.md';
+import Readme from '../README.md';
 import DateTimeInput from './date-time-input';
-import Section from '../../../../.storybook/decorators/section';
+import Section from '../../../../../.storybook/decorators/section';
 
 class DateTimeInputStory extends React.Component {
   static displayName = 'DateTimeInputStory';
