+++ packages/components/inputs/date-time-input/src/date-time-input.visualroute.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { DateTimeInput } from '@commercetools-frontend/ui-kit';
-import { Suite, Spec } from '../../../../test/percy';
+import { Suite, Spec } from '../../../../../test/percy';
 
 const value = '2018-11-13 15:00';
