+++ packages/components/inputs/date-time-input/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './date-time-input';
+
+export { default as version } from './version';
