--- src/components/inputs/date-time-input/index.js
@@ -1,1 +1,0 @@
-export { default } from './date-time-input';
